#38
- config: реструктуризация раздела
